signer/requirements.txt
x10-python-trading-starknet==1.3.1
flask
gunicorn
